requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via griffe
distro==1.9.0
    # via openai
fastapi==0.115.12
    # via -r requirements.in
griffe==1.7.1
    # via openai-agents
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   mcp
    #   openai
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
mcp==1.6.0
    # via
    #   -r requirements.in
    #   openai-agents
openai==1.69.0
    # via
    #   -r requirements.in
    #   openai-agents
openai-agents==0.0.7
    # via -r requirements.in
pydantic==2.11.1
    # via
    #   -r requirements.in
    #   fastapi
    #   mcp
    #   openai
    #   openai-agents
    #   pydantic-settings
pydantic-core==2.33.0
    # via pydantic
pydantic-settings==2.8.1
    # via mcp
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.20
    # via -r requirements.in
requests==2.32.3
    # via openai-agents
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sse-starlette==2.2.1
    # via mcp
starlette==0.46.1
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
tqdm==4.67.1
    # via openai
types-requests==2.32.0.20250328
    # via openai-agents
typing-extensions==4.13.0
    # via
    #   fastapi
    #   openai
    #   openai-agents
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via
    #   requests
    #   types-requests
uvicorn==0.34.0
    # via
    #   -r requirements.in
    #   mcp
